[ADD] Support for bigint columns
Mark models with _bigint_id = True or set bigint=True in fields.Integer

+++ openerp/addons/base/base.sql
@@ -125,7 +125,7 @@ CREATE TABLE ir_model_data (
     date_update timestamp without time zone,
     module varchar NOT NULL,
     model varchar NOT NULL,
-    res_id integer,
+    res_id bigint,
     primary key(id)
 );
